fixing some memory leaks and adding progress report

+++ src/analysis/E4NuAnalysis.h
@@ -30,12 +30,12 @@ namespace e4nu {
     e4nu::EventI * GetValidEvent( const unsigned int event_id ) ;
     unsigned int GetNEvents( void ) const ;
 
-    unsigned int fNEventsAfterEMomCut = 0 ; 
-    unsigned int fNEventsAfterPhiCut = 0 ; 
-    unsigned int fNEventsAfterQ2Cut = 0 ; 
-    unsigned int fNEventsAfterWCut = 0 ; 
-    unsigned int fNEventsAfterPhiOpeningAngleCut = 0 ; 
-    unsigned int fNEventsAfterThetaCut = 0 ;     
+    long int fNEventsAfterEMomCut = 0 ; 
+    long int fNEventsAfterPhiCut = 0 ; 
+    long int fNEventsAfterQ2Cut = 0 ; 
+    long int fNEventsAfterWCut = 0 ; 
+    long int fNEventsAfterPhiOpeningAngleCut = 0 ; 
+    long int fNEventsAfterThetaCut = 0 ;     
 
     virtual ~E4NuAnalysis();
